+++ src/QuickMarcEditor/QuickMarcEditor.js
@@ -137,6 +137,14 @@ const QuickMarcEditor = ({
   const searchParameters = new URLSearchParams(location.search);
   const isShared = searchParameters.get('shared') === 'true';
 
+  const saveLastFocusedInput = useCallback((e) => {
+    lastFocusedInput.current = e.target;
+  }, [lastFocusedInput]);
+
+  const focusLastFocusedInput = useCallback(() => {
+    lastFocusedInput.current?.focus();
+  }, [lastFocusedInput]);
+
   const { unlinkAuthority } = useAuthorityLinking({ marcType, action });
 
   useEffect(() => {
@@ -156,9 +164,9 @@ const QuickMarcEditor = ({
   const handleSubmitResponse = useCallback(() => {
     if (continueAfterSave.current) {
       continueAfterSave.current = false;
-      lastFocusedInput.current?.focus();
+      focusLastFocusedInput();
     }
-  }, [continueAfterSave]);
+  }, [continueAfterSave, focusLastFocusedInput]);
 
   const closeModals = () => {
     setIsDeleteModalOpened(false);
@@ -208,7 +216,7 @@ const QuickMarcEditor = ({
     };
   }, [setIsValidationModalOpen, isBackEndValidationMarcType, marcType]);
 
-  const showValidationIssuesCallouts = useCallback((issues) => {
+  const showErrorsForMissingFields = useCallback((issues) => {
     issues.forEach((error) => {
       showCallout({
         message: error.message,
@@ -216,6 +224,33 @@ const QuickMarcEditor = ({
         values: error.values,
         type: error.severity === SEVERITY.ERROR ? 'error' : 'warning',
       });
+    });
+  }, [showCallout]);
+
+  const showValidationIssuesToasts = useCallback((validationErrors) => {
+    const allIssuesArray = Object.values(validationErrors).flat();
+    const failCount = allIssuesArray.filter(issue => issue.severity === SEVERITY.ERROR).length;
+    const warnCount = allIssuesArray.length - failCount;
+
+    const values = {
+      warnCount,
+      failCount,
+      breakingLine: <br />,
+    };
+    let messageId = null;
+
+    if (failCount && warnCount) {
+      messageId = 'ui-quick-marc.record.save.error.failAndWarn';
+    } else if (failCount) {
+      messageId = 'ui-quick-marc.record.save.error.fail';
+    } else {
+      messageId = 'ui-quick-marc.record.save.error.warn';
+    }
+
+    showCallout({
+      messageId,
+      values,
+      type: failCount ? 'error' : 'warning',
     });
   }, [showCallout]);
 
@@ -252,9 +287,12 @@ const QuickMarcEditor = ({
 
         return;
       }
+
       const validationErrorsWithoutFieldId = newValidationErrors[MISSING_FIELD_ID] || [];
 
-      showValidationIssuesCallouts(validationErrorsWithoutFieldId);
+      showErrorsForMissingFields(validationErrorsWithoutFieldId);
+      showValidationIssuesToasts(newValidationErrors);
+      focusLastFocusedInput();
       setIsValidatedCurrentValues(true);
     } else {
       setValidationErrors({});
@@ -276,13 +314,15 @@ const QuickMarcEditor = ({
     getState,
     hasErrorIssues,
     setValidationErrors,
-    showValidationIssuesCallouts,
+    showErrorsForMissingFields,
     showCallout,
     validate,
     runConfirmationChecks,
     isValidatedCurrentValues,
     setIsValidatedCurrentValues,
     manageBackendValidationModal,
+    focusLastFocusedInput,
+    showValidationIssuesToasts,
     continueAfterSave,
   ]);
 
@@ -443,10 +483,6 @@ const QuickMarcEditor = ({
       setRecords(values.records);
     }
   }, []);
-
-  const saveLastFocusedInput = useCallback((e) => {
-    lastFocusedInput.current = e.target;
-  }, [lastFocusedInput]);
 
   const shortcuts = useMemo(() => ([{
     name: 'save',